Daisy Check Block
This is the building block of the daisy chain that takes an interrupt request from a peripheral device (IRQ_input) and an interrupt acknowledgement signal from the CPU (Ack_input) as inputs and determines whether to pass the Ack_input to the next block when IRQ_input is inactive or to the peripheral device when IRQ_input is active. 
The two outputs are dedicated to pass the Ack_input to the next block or to the peripheral device. IRQ_input is passed over to the interrupt input of the CPU.